requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
distlib==0.3.4
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
filelock==3.6.0
ftd==0.1.0
pipenv==2022.3.24
platformdirs==2.5.1
psycopg2==2.9.3
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
virtualenv==20.14.0
virtualenv-clone==0.5.7
